+++ sdmx-dl-provider-ri/src/main/java/internal/util/http/ext/DumpingClient.java
@@ -3,9 +3,9 @@ package internal.util.http.ext;
 import internal.util.http.HttpClient;
 import internal.util.http.HttpRequest;
 import internal.util.http.HttpResponse;
-import sdmxdl.format.MediaType;
 import lombok.NonNull;
 import nbbrd.io.Resource;
+import sdmxdl.format.MediaType;
 
 import java.io.IOException;
 import java.io.InputStream;